LOG-5725: Refactor multilineexception functional test
