+++ test/functional/filters/multilineexception/suite_test.go
@@ -1,0 +1,13 @@
+package multilineexception
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestSuite(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "[functional][filters][multilineerror]")
+}
